requirements.txt
python-telegram-bot[webhooks]>=21.0,<22.0
supabase>=2.0.0
APScheduler>=3.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
